coyote is now not a workspace
