--- component_string/Cargo.toml
@@ -1,12 +1,0 @@
-[package]
-name = "component_string"
-version = "0.1.0"
-edition = "2021"
-
-# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
-
-[dependencies]
-rulesets = { path = "../rulesets" }
-parse = { path = "../parse" }
-coyote = { path = "../coyote" }
-template_steps = { path = "../template_steps" }
