--- coyote/tests/function_components.rs
@@ -1,20 +1,0 @@
-use coyote::{attr_val, list, text, tmpl, Component};
-
-// Test will not build if Function Components do not build
-
-fn woof() -> Component {
-    tmpl("<input type=submit value=\"yus -_-\">", [])
-}
-
-fn woof_woof() -> Component {
-    let descendants = list([text("you're a boy kisser aren't you >:3"), woof()]);
-
-    let attributes = list([attr_val("action", "/uwu"), attr_val("method", "post")]);
-
-    tmpl("<form {}>{}</form>", [attributes, descendants])
-}
-
-#[test]
-fn test_coyote_api() {
-    let _woof_form = woof_woof();
-}
