--- coyote_html/Cargo.toml
@@ -1,12 +1,0 @@
-[package]
-name = "coyote_html"
-version = "0.1.0"
-edition = "2021"
-
-# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
-
-[dependencies]
-rulesets = { path = "../rulesets" }
-coyote = { path = "../coyote" }
-template_steps = { path = "../template_steps" }
-component_string = { path = "../component_string" }
