--- rulesets/Cargo.toml
@@ -1,6 +1,0 @@
-[package]
-name = "rulesets"
-version = "0.1.0"
-edition = "2021"
-
-[dependencies]
