+++ src/lib.rs
@@ -8,7 +8,8 @@ mod sliding_window;
 mod tag_info;
 mod template_steps;
 
-pub mod html;
+mod html;
+mod xml;
 
 pub use crate::coyote::attr;
 pub use crate::coyote::attr_val;
@@ -18,5 +19,8 @@ pub use crate::coyote::tmpl;
 pub use crate::coyote::unescaped_text;
 pub use crate::coyote::vlist;
 pub use crate::coyote::Component;
+
+pub use crate::html::{ClientHtml, Html};
+pub use crate::xml::Xml;
 
 // pub use crate::html;
