--- template_steps/Cargo.toml
@@ -1,11 +1,0 @@
-[package]
-name = "template_steps"
-version = "0.1.0"
-edition = "2021"
-
-# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
-
-[dependencies]
-rulesets = { path = "../rulesets" }
-parse = { path = "../parse" }
-coyote = { path = "../coyote" }
