+++ tests/function_components.rs
@@ -1,6 +1,6 @@
 use coyote::{attr_val, list, text, tmpl, Component};
 
-// Test will not build if Function Components do not build
+// Test will not build if function components do not build
 
 fn woof() -> Component {
     tmpl("<input type=submit value=\"yus -_-\">", [])
