+++ tests/html_arc_mutex_tests.rs
@@ -1,5 +1,5 @@
-use coyote::html::Html;
-use coyote::{attr_val, list, text, tmpl, Component};
+// use coyote::html::Html;
+use coyote::{attr_val, list, text, tmpl, Component, Html};
 use std::sync::Arc;
 use std::sync::Mutex;
