+++ tests/pretty_html_tests.rs
@@ -1,6 +1,4 @@
-// use html::compose;
-use coyote::html::{ClientHtml, Html};
-use coyote::tmpl;
+use coyote::{tmpl, ClientHtml, Html};
 
 #[test]
 fn test_pretty_html_no_empty_space() {
